Extract directives to own files
