+++ libs/components/src/button/button.module.ts
@@ -1,11 +1,26 @@
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 
-import { ButtonComponent, ButtonPrefixDirective, ButtonPrimaryDirective } from "./button.component";
+import { ButtonComponent } from "./button.component";
+import {
+  ButtonPrimaryDirective,
+  ButtonSecondaryDirective,
+  ButtonDangerDirective,
+} from "./button-style.directive";
 
 @NgModule({
   imports: [CommonModule],
-  exports: [ButtonComponent, ButtonPrimaryDirective, ButtonPrefixDirective], // etc
-  declarations: [ButtonComponent, ButtonPrimaryDirective, ButtonPrefixDirective],
+  exports: [
+    ButtonComponent,
+    ButtonPrimaryDirective,
+    ButtonSecondaryDirective,
+    ButtonDangerDirective,
+  ],
+  declarations: [
+    ButtonComponent,
+    ButtonPrimaryDirective,
+    ButtonSecondaryDirective,
+    ButtonDangerDirective,
+  ],
 })
 export class ButtonModule {}
