+++ libs/components/src/icon-button/icon-button.module.ts
@@ -2,10 +2,34 @@ import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 
 import { BitIconButtonComponent } from "./icon-button.component";
+import {
+  IconButtonMainDirective,
+  IconButtonMutedDirective,
+  IconButtonDangerDirective,
+  IconButtonPrimaryDirective,
+  IconButtonContrastDirective,
+  IconButtonSecondaryDirective,
+} from "./icon-button-styles.directive";
 
 @NgModule({
   imports: [CommonModule],
-  declarations: [BitIconButtonComponent],
-  exports: [BitIconButtonComponent],
+  declarations: [
+    BitIconButtonComponent,
+    IconButtonMainDirective,
+    IconButtonMutedDirective,
+    IconButtonDangerDirective,
+    IconButtonPrimaryDirective,
+    IconButtonContrastDirective,
+    IconButtonSecondaryDirective,
+  ],
+  exports: [
+    BitIconButtonComponent,
+    IconButtonMainDirective,
+    IconButtonMutedDirective,
+    IconButtonDangerDirective,
+    IconButtonPrimaryDirective,
+    IconButtonContrastDirective,
+    IconButtonSecondaryDirective,
+  ],
 })
 export class IconButtonModule {}
